+++ cloud-dev-project/models/index.ts
@@ -2,7 +2,6 @@ export * from './loginModel';
 export * from './homePageModel';
 export * from './Profile';
 export * from './BlogListModel';
-export * from './EditProfile';
 export * from './BlogListModel';
 export * from './CreateBlogModel';
